requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asari==0.2.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.0.0
Babel==2.15.0
bce-python-sdk==0.9.7
beartype==0.18.5
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
coverage==5.5
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
datasets==2.18.0
decorator==4.4.2
deprecation==2.1.0
dill==0.3.8
dnspython==2.6.1
editdistance==0.8.1
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
eyed3==0.9.7
fastapi==0.111.0
fastapi-cli==0.0.2
ffmpy==0.3.2
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
funasr==1.0.26
future==1.0.0
gast==0.5.4
google==3.0.0
gradio==3.50.0
gradio_client==0.6.1
h11==0.14.0
hmmlearn==0.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jaconv==0.3.4
jamo==0.4.1
Janome==0.4.2
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.9
modelscope==1.14.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
onnxconverter-common==1.14.0
onnxruntime==1.17.3
openai-whisper==20231117
opencv-python==4.9.0.80
orjson==3.10.2
oss2==2.18.5
packaging==21.3
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
pocketsphinx==5.0.3
pooch==1.8.1
pretty-errors==1.2.25
proglog==0.1.10
protobuf==3.20.2
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyAudioAnalysis==0.3.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pymlask @ git+https://github.com/ikegami-yukino/pymlask@227ce3b42bba9ede71165ec10735ed9d18f6aa37
pynndescent==0.5.12
pyparsing==3.1.2
pysubs2==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-wpe==0.0.1
pytz==2024.1
PyYAML==6.0.1
rarfile==4.2
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
resampy==0.2.2
rich==13.7.1
rotary-embedding-torch==0.6.1
rpds-py==0.18.0
ruff==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
skl2onnx==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.3
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torch-complex==0.4.3
torchaudio==2.3.0
torchinfo==1.8.0
torchvision==0.18.0
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
umap-learn==0.5.6
unidic-lite==1.0.8
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
visualdl==2.5.3
watchfiles==0.21.0
websockets==11.0.3
Werkzeug==3.0.3
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
